Refactor12 - stable release (#26)
* add modifier checkpoints

add modifier checkpoints in addition to checksum checkpoints

* use height not time for mod checkpoints

* some sync fixes

* uncomment some sync fixes that shouldnt have been commented out

* fixed issue where it would look for nonexistant blockindex when computing initial stake modifier

* PoS changes

PoS changes

* fix PoW check

* remove work check from header validation. we dont have enough info at the time to evaluate for this

* edited stake modifier creation

* further edits to how we get next stake modifier

further edits to how we get next stake modifier

* adjust computing next stake modifier for PoW blocks

adjust computing next stake modifier for PoW blocks

* removed bad else statement

removed bad else statement

* change stake modifier type

change stake modifier type

* add backwards compatibility for sync start

add backwards compatibility for sync start so that we can sync from an
older node that doesnt have header messages

* fix client name

fix client name

* only connect to newest version

only connect to newest version

* a few sync and staking fixes. staking isnt completely fixed though

a few sync and staking fixes. staking isnt completely fixed though

* lower transition block

lower transition block

* fix header sync issue with bytes not lining up

fix header sync issue with bytes not lining up because there was no skip
when reading the stream to account for the vchBlockSig

* see headers in getinfo

see headers in getinfo

* some staking adjustments

some staking adjustments

* fixed staking errors

fixed staking errors

* fixed remaining hashTarget

fixed remaining hashTarget

* some changes to staking

some changes to staking to correct height for when we are making a new
stake, we want to calculate with the height the block is going to be at
not the one we are currently on

* use pos limit for pos not pow limit

use pos limit for pos not pow limit

* some pos algo changes (not done)

* should be 20 bits not 5

mistake when shifting bits. it was set to 5, but it is 5 hex places and
each hex is 4 bits so it should be 20 bits

* final staking kernel fixes

final staking kernel fixes

* 6 because coin is only 6 to the right of the decimal

* should parse for 6 decimalss, not 8

should parse for 6 decimalss, not 8